pyproject.toml
[tool.poetry]
name = "mac-rename"
version = "0.1.0"
description = ""
authors = ["Moritz Enderle <contact@moritzenderle.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12 < 3.13"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-multipart = "^0.0.6"
aiofiles = "^23.2.1"
numpy = "^2.1.3"
rawpy = "^0.23.2"
watchdog = "^6.0.0"
tqdm = "^4.67.1"
rich = "^14.0.0"
pillow = "^10.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
